Refactor mappers

+++ lib/Db/AccessTokenMapper.php
@@ -100,8 +100,7 @@ class AccessTokenMapper extends Mapper {
 			throw new InvalidArgumentException('Argument client_id must be an int and user_id must be a string');
 		}
 
-		$sql = 'DELETE FROM `' . $this->tableName . '` '
-			. 'WHERE client_id = ? AND user_id = ?';
+		$sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ? AND `user_id` = ?';
 		$stmt = $this->execute($sql, [$clientId, $userId], null, null);
 		$stmt->closeCursor();
 	}
@@ -123,15 +122,13 @@ class AccessTokenMapper extends Mapper {
      * @param int $clientId The client ID
      * @see SettingsController::deleteClient()
      */
-	public function deleteByClient($clientId)
-    {
+	public function deleteByClient($clientId) {
         if (!is_int($clientId)) {
             throw new InvalidArgumentException('Argument client_id must be an int');
         }
 
-        $sql = 'DELETE FROM `' . $this->tableName . '`'
-            . 'WHERE client_id = ?';
-        $stmt = $this->execute($sql, array($clientId), null, null);
+        $sql = 'DELETE FROM `' . $this->tableName . '` ' . ' WHERE `client_id` = ?';
+        $stmt = $this->execute($sql, [$clientId], null, null);
         $stmt->closeCursor();
     }
 }
