+++ lib/Db/AuthorizationCodeMapper.php
@@ -99,8 +99,7 @@ class AuthorizationCodeMapper extends Mapper {
 			throw new InvalidArgumentException('Argument client_id must be an int and user_id must be a string');
 		}
 
-		$sql = 'DELETE FROM `' . $this->tableName . '` '
-			. 'WHERE client_id = ? AND user_id = ?';
+		$sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ? AND `user_id` = ?';
 		$stmt = $this->execute($sql, [$clientId, $userId], null, null);
 		$stmt->closeCursor();
 	}
@@ -108,7 +107,7 @@ class AuthorizationCodeMapper extends Mapper {
 	/**
 	 * Deletes all entities from the table
 	 */
-	public function deleteAll(){
+	public function deleteAll() {
 		$sql = 'DELETE FROM `' . $this->tableName . '`';
 		$stmt = $this->execute($sql, []);
 		$stmt->closeCursor();
@@ -122,15 +121,13 @@ class AuthorizationCodeMapper extends Mapper {
      * @param int $clientId The client ID
      * @see SettingsController::deleteClient()
      */
-    public function deleteByClient($clientId)
-    {
+    public function deleteByClient($clientId) {
         if (!is_int($clientId)) {
             throw new InvalidArgumentException('Argument client_id must be an int');
         }
 
-        $sql = 'DELETE FROM `' . $this->tableName . '`'
-            . 'WHERE client_id = ?';
-        $stmt = $this->execute($sql, array($clientId), null, null);
+        $sql = 'DELETE FROM `' . $this->tableName . '` ' . 'WHERE `client_id` = ?';
+        $stmt = $this->execute($sql, [$clientId], null, null);
         $stmt->closeCursor();
     }
